backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pdfminer.six==20231228
httpx==0.25.2
python-json-logger==2.0.7